feat-JSMP-3: add INIT action
