+++ src/app/redux/reducers/authReducer.ts
@@ -1,16 +1,11 @@
 // actions
 import {
-  // LOGIN_REQUEST,
-  // SINGUP_REQUEST,
   AUTH_SUCCESS,
-  // AUTH_FAILURE,
   LOGOUT,
 } from "../actions";
 
 // types
 import { Action } from "../actions/types";
-
-// type for initial state and export
 
 const initialState = {
   user: null,
@@ -21,7 +16,7 @@ export default function authReducer(state = initialState, { type, payload }: Act
     case AUTH_SUCCESS:
       return { ...state, user: payload };
     case LOGOUT:
-      return { ...state, user: null };
+      return { ...state, user: null };  
     default:
       return state;
   }
