+++ src/app/redux/sagas/authSaga.ts
@@ -8,36 +8,30 @@ import { Creds } from "../../services/types";
 import {
   LOGIN_REQUEST,
   SINGUP_REQUEST,
-  AUTH_SUCCESS,
   AUTH_FAILURE,
   LOGOUT,
+  authSuccess,
+  INIT,
 } from "../actions";
 
 // auth services
-import { signIn, signUp, logOut } from "../../services/auth";
+import * as authService from "../../services/auth";
 
-const authService: {
-  signIn: typeof signIn;
-  signUp: typeof signUp;
-  logOut: typeof logOut;
-} = {
-  signIn,
-  signUp,
-  logOut,
-};
+type AuthService = typeof authService;
 
-type AuthMethod = keyof typeof authService;
+type AuthMethod = keyof AuthService;
 
 const typeToMethodMap: { [key: string]: AuthMethod } = {
   [LOGIN_REQUEST]: "signIn",
   [SINGUP_REQUEST]: "signUp",
-  [LOGOUT]: "logOut",
+  [LOGOUT]: "logOut",  
+  [INIT]: "init",
 };
 
 // worker sagas
 export const authUser = async (creds: Creds, authType: AuthMethod) => {
   try {
-    const userData = await authService[authType]?.(creds);   
+    const userData = await authService[authType]?.(creds);
     return userData;
   } catch (err) {
     console.log("auth error:", err);
@@ -51,12 +45,9 @@ export function* authHandler({ type, payload }: Action): Generator<any> {
     const result = yield call(authService[authType]);
     console.log(result);
   } else {
-    const userData = yield call(authUser, payload as Creds, authType);
+    const userData: any = yield call(authUser, payload as Creds, authType);
     if (userData) {
-      yield put({
-        type: AUTH_SUCCESS,
-        // payload: {}
-      });
+      yield put(authSuccess(userData.user.email));
     } else {
       yield put({
         type: AUTH_FAILURE,
@@ -65,7 +56,12 @@ export function* authHandler({ type, payload }: Action): Generator<any> {
   }
 }
 
+export function* initHandler({ type, payload }: Action): Generator<any> {
+
+}
+
 // watcher saga
 export default function* authSaga() {
   yield takeLatest([LOGIN_REQUEST, SINGUP_REQUEST, LOGOUT], authHandler);
+  yield takeLatest([INIT], initHandler);
 }
